add support for derivative metrics
A derivative metric is like a meter but accepts an absolute counter as
input. This is useful for metrics like bytes sent over the network or
cpu cycles which are generally monotonically increasing counters and
therefore need to be derived from the previous sample to get a useful
rate/s value.