Add README example of timer with output.
It's pretty cool that metrik's default timer implementation has a much better UX than ruby's native Benchmark suite: it returns the output of the block, rather than a report. This makes it much easier to drop into various spots in an existing codebase without having to change much.

I had to go digging through the docs and then the source to confirm it though, so I thought an upfront example would help!